Wrote vector implementations for comparison I was interested in seeing what the performance difference would be between generic descriptive methods based on lists vs methods tailored to work specifically on vectors of doubles (provided by the vector package).
I expected `Vector` to come out ahead, but was still surprised at how
far it blew lists out of the water:

| Method | Average runtime for 1,000,000 elements |
|--------|---------------------------------------:|
| mean    | 100.1 ms |
| meanVec | 3.6 ms |
| var     | 675.9 ms |
| varVec  | 125.0 ms |
| std     | 673.3 ms |
| stdVec  | 125.0 ms |

From these results I've decided to let `Vector` based implementations be
the default, and only provide generics where explicitly needed. After
all, a stated goal of this project is to outperform implementations in
other languages; I see this as a step in that direction.