Added a function for binning continuous data Generates fixed number of even-width bins. Might move it to a different module.
The complexity looks pretty bad, but it's from the `n log n` step in
calling `percentile`. This is because `percentile` needs a sorted list,
but doesn't assume that the argument is sorted. After the sorting step,
`percentile` is O(1), so I might look at doing an implementation that
assumes the input list is sorted.